+++ tests/Stratara.Projections.Tests/Services/ProjectionWorkerIntegrityTests.cs
@@ -155,6 +155,7 @@ public class ProjectionWorkerIntegrityTests
                 PipelineProvider.Object,
                 Options.Create(new BusEnvelopeJsonOptions { MaxBodyBytes = maxBodyBytes }),
                 Options.Create(new BusEnvelopeIntegrityOptions { Mode = integrityMode }),
+                Options.Create(new ProjectionOptions()),
                 signer);
         }
     }